refactor: split 3 monolithic scripts into focused modules (issue #744)
- mypy_autofix_agent.py (→262 LOC): extract data types, constants, IO
  helpers, and all 5 fix strategies into mypy_fixers.py; agent is now a
  thin orchestrator that imports from mypy_fixers
- run_assessment.py (→163 LOC): extract 15 per-category analysis
  functions into assessment_modules.py with dispatch table; orchestrator
  retains wrappers, scoring, and report generation
- combat_manager.py (→231 LOC): extract explosion helpers (bomb, laser,
  generic AOE) into _CombatExplosionsMixin (combat_explosions.py) and
  shot visual helpers into _CombatVisualsMixin (combat_visuals.py);
  CombatManagerBase inherits from both; all public APIs preserved
- Update test patch targets to scripts.assessment_modules.*
- Bump SPEC.md to 1.1.31
