+++ SPEC.md
@@ -409,6 +409,7 @@ Active development. Core games (F1-F6, F8-F9) fully implemented and tested. F7 (
 
 | Date       | Version | Changes                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         |
 | ---------- | ------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| 2026-04-14 | 1.1.31  | Refactor(#744): split 3 monolithic scripts into focused modules — `mypy_autofix_agent.py` (731→262 LOC) extracts fix strategies and data types into `mypy_fixers.py`; `run_assessment.py` (628→163 LOC) extracts 15 per-category analysis functions into `assessment_modules.py`; `combat_manager.py` (624→231 LOC) extracts explosion helpers into `combat_explosions.py` and shot visual helpers into `combat_visuals.py`. All public APIs preserved.                                                                                         |
 | 2026-04-14 | 1.1.28  | Raised statement coverage toward 80% by adding 8 test files (126+ tests): Zombie Survival atmosphere_manager, weapon_system, screen_event_handler, renderer, and game delegation; Duum renderer and screen_event_handler extensions; Force Field combat_actions. Addresses issue #738.                                                                                                                                                                                                                                                          |
 | 2026-04-14 | 1.1.26  | Removed tracked coverage artifact directories (`cov_test3/`, `cov_annot2/`, `cov.json`) from git history and added coverage output patterns to `.gitignore` to prevent future accidental commits. Addresses issue #766.                                                                                                                                                                                                                                                                                                                         |
 | 2026-04-14 | 1.1.25  | Extracted `raycaster_textures` module from the monolithic `raycaster.py` to reduce file size and improve separation of concerns; decomposed oversized rendering functions into focused single-responsibility helpers. Addresses issue #741.                                                                                                                                                                                                                                                                                                     |
