+++ scripts/run_assessment.py
@@ -1,22 +1,23 @@
 #!/usr/bin/env python3
-"""
-Run a specific assessment (A-O) on the repository.
+"""Run a specific assessment (A-O) on the repository.
 
-This script executes an individual assessment and generates a detailed report
-based on actual code analysis.
+Per-category analysis logic lives in assessment_modules.py.
+This file is the thin orchestrator: wrappers, scoring, report generation.
 """
+
+from __future__ import annotations
 
 import argparse
 import os
-import re
 import sys
 from datetime import datetime
 from pathlib import Path
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from scripts.shared.logging_config import setup_script_logging
-from scripts.shared.subprocess_utils import (
+from scripts.assessment_modules import analyze_module, file_discovery  # noqa: E402
+from scripts.shared.logging_config import setup_script_logging  # noqa: E402
+from scripts.shared.subprocess_utils import (  # noqa: E402,F401
     run_black_check,
     run_command,
     run_ruff_check,
@@ -24,7 +25,6 @@ from scripts.shared.subprocess_utils import (
 
 logger = setup_script_logging()
 
-# Assessment definitions
 ASSESSMENTS = {
     "A": {"name": "Code Structure", "description": "Code structure and organization"},
     "B": {"name": "Documentation", "description": "README, docstrings, comments"},
@@ -42,26 +42,6 @@ ASSESSMENTS = {
     "N": {"name": "Scalability", "description": "Performance at scale"},
     "O": {"name": "Maintainability", "description": "Code maintainability"},
 }
-
-# Directories excluded from Python file discovery
-_EXCLUDED_DIRS = {
-    ".git",
-    "__pycache__",
-    ".venv",
-    "venv",
-    "node_modules",
-    ".tox",
-    "build",
-    "dist",
-}
-
-
-def find_python_files() -> list[Path]:
-    """Find all Python files in the repository."""
-    python_files: list[Path] = []
-    for pattern in ["**/*.py"]:
-        python_files.extend(Path(".").glob(pattern))
-    return [f for f in python_files if not any(p in f.parts for p in _EXCLUDED_DIRS)]
 
 
 def run_ruff_check_wrapper() -> dict:
@@ -89,419 +69,8 @@ def run_black_check_wrapper() -> dict:
         return {"exit_code": -1, "files_to_format": 0, "errors": "black not installed"}
 
 
-def count_test_files() -> int:
-    """Count test files in the repository."""
-    test_patterns = ["**/test_*.py", "**/*_test.py", "**/tests/*.py"]
-    test_files: set[Path] = set()
-    for pattern in test_patterns:
-        test_files.update(Path(".").glob(pattern))
-    return len(test_files)
-
-
-def check_documentation() -> dict:
-    """Check documentation status."""
-    has_readme = Path("README.md").exists()
-    has_docs = Path("docs").exists()
-    has_changelog = Path("CHANGELOG.md").exists()
-    return {
-        "has_readme": has_readme,
-        "has_docs_dir": has_docs,
-        "has_changelog": has_changelog,
-    }
-
-
-def grep_in_files(pattern: str, files: list[Path]) -> int:
-    """Count files containing a pattern."""
-    count = 0
-    for file in files:
-        try:
-            content = file.read_text(encoding="utf-8")
-            if re.search(pattern, content):
-                count += 1
-        except Exception:  # noqa: BLE001  # noqa: BLE001
-            pass
-    return count
-
-
-# ---------------------------------------------------------------------------
-# Extracted single-responsibility helpers
-# ---------------------------------------------------------------------------
-
-
-def file_discovery() -> tuple[list[Path], int]:
-    """Discover Python source files and return them with a count.
-
-    Returns:
-        A tuple of (python_files, file_count) where python_files is the list
-        of discovered Path objects and file_count is len(python_files).
-    """
-    python_files = find_python_files()
-    return python_files, len(python_files)
-
-
-def analyze_module(
-    assessment_id: str,
-    python_files: list[Path],
-    file_count: int,
-) -> tuple[list[str], int | None]:
-    """Analyse a repository for a single assessment category.
-
-    Each assessment category maps to a focused analysis routine.  The function
-    collects human-readable *findings* lines and an integer *score* (0-10,
-    or None when the score cannot be determined automatically).
-
-    Args:
-        assessment_id: Single uppercase letter identifying the assessment (A-O).
-        python_files:  List of Python source Path objects from file_discovery().
-        file_count:    Number of entries in python_files.
-
-    Returns:
-        A tuple of (findings, score) where findings is a list of markdown
-        bullet strings and score is an int 0-10 or None.
-    """
-    findings: list[str] = []
-    score: int | None = 10  # Start with perfect score
-
-    if assessment_id == "A":  # Code Structure
-        has_src = Path("src").exists()
-        has_tests = (
-            Path("tests").exists()
-            or Path("src/tests").exists()
-            or (file_count > 0 and count_test_files() > 0)
-        )
-        findings.append(f"- Python files found: {file_count}")
-        findings.append(
-            f"- Source directory structure (src/): {'✓' if has_src else '✗'}"
-        )
-        findings.append(f"- Tests directory/files: {'✓' if has_tests else '✗'}")
-        if not has_src:
-            score -= 2  # type: ignore[operator]
-        if not has_tests:
-            score -= 2  # type: ignore[operator]
-
-    elif assessment_id == "B":  # Documentation
-        docs = check_documentation()
-        findings.append(f"- README.md: {'✓' if docs['has_readme'] else '✗'}")
-        findings.append(f"- docs/ directory: {'✓' if docs['has_docs_dir'] else '✗'}")
-        findings.append(f"- CHANGELOG.md: {'✓' if docs['has_changelog'] else '✗'}")
-
-        # Check for docstrings in top 10 files
-        docstring_count = 0
-        checked_files = 0
-        for p in python_files[:10]:
-            content = p.read_text(encoding="utf-8")
-            if '"""' in content or "'''" in content:
-                docstring_count += 1
-            checked_files += 1
-
-        if checked_files > 0:
-            docstring_percent = (docstring_count / checked_files) * 100
-            findings.append(f"- Docstring presence (sample): {docstring_percent:.0f}%")
-            if docstring_percent < 50:
-                score -= 2  # type: ignore[operator]
-
-        if not docs["has_readme"]:
-            score -= 3  # type: ignore[operator]
-        if not docs["has_docs_dir"]:
-            score -= 1  # type: ignore[operator]
-
-    elif assessment_id == "C":  # Test Coverage
-        test_count = count_test_files()
-        findings.append(f"- Test files found: {test_count}")
-        if test_count == 0:
-            score -= 5  # type: ignore[operator]
-            findings.append("- Critical: No tests found")
-        elif test_count < 5:
-            score -= 2  # type: ignore[operator]
-            findings.append("- Warning: Low number of test files")
-        else:
-            findings.append("- Good number of test files")
-
-        # Attempt to run pytest with coverage
-        try:
-            # We don't fail if this errors, as environment might not support it
-            env = os.environ.copy()
-            env["PYTHONPATH"] = "."
-            cmd = [
-                sys.executable,
-                "-m",
-                "pytest",
-                "--cov=src",
-                "--cov-report=term-missing",
-            ]
-            result = run_command(cmd, check=False, env=env)
-            if result.returncode == 0:
-                findings.append("- Pytest execution: ✓ Passed")
-                # Parse coverage percentage if possible (simple heuristic)
-                match = re.search(r"TOTAL\s+\d+\s+\d+\s+(\d+)%", result.stdout)
-                if match:
-                    cov_percent = int(match.group(1))
-                    findings.append(f"- Coverage: {cov_percent}%")
-                    if cov_percent < 50:
-                        score -= 2  # type: ignore[operator]
-                        findings.append("- Warning: Low test coverage (< 50%)")
-                    elif cov_percent < 80:
-                        score -= 1  # type: ignore[operator]
-                        findings.append("- Note: Moderate test coverage (< 80%)")
-            else:
-                findings.append(
-                    f"- Pytest execution: ✗ Failed (Exit code {result.returncode})"
-                )
-                score -= 2  # type: ignore[operator]
-        except Exception as e:  # noqa: BLE001
-            findings.append(f"- Pytest execution failed to start: {e}")
-
-    elif assessment_id == "D":  # Error Handling
-        try_count = grep_in_files(r"try:", python_files)
-        except_pat = r"except Exception as e:"  # noqa: BLE001
-        except_count = grep_in_files(except_pat, python_files)
-        findings.append(f"- Files with try blocks: {try_count}")
-        findings.append(f"- Files with except blocks: {except_count}")
-        if try_count == 0:
-            score -= 2  # type: ignore[operator]
-            findings.append("- Warning: No error handling detected in sample")
-
-    elif assessment_id == "E":  # Performance
-        # Check for profilers or common optimization patterns
-        perf_imports = grep_in_files(
-            r"import cProfile|import timeit|import pstats", python_files
-        )
-        findings.append(f"- Files with performance tools: {perf_imports}")
-
-        # Check for time.sleep which is often an anti-pattern
-        time_sleep = grep_in_files(r"time\.sleep", python_files)
-        findings.append(f"- Files using time.sleep: {time_sleep}")
-
-        if perf_imports == 0:
-            findings.append(
-                "- Note: No explicit performance profiling tools found in code"
-            )
-            score -= 1  # type: ignore[operator]
-
-        if time_sleep > 2:
-            findings.append(
-                "- Warning: Multiple uses of time.sleep detected (potential bottleneck)"
-            )
-            score -= 1  # type: ignore[operator]
-
-    elif assessment_id == "F":  # Security
-        # Basic check for subprocess without shell=True or hardcoded secrets
-        subprocess_shell = grep_in_files(r"subprocess\..*shell=True", python_files)
-        if subprocess_shell > 0:
-            findings.append(
-                f"- Critical: subprocess with shell=True found in "
-                f"{subprocess_shell} files"
-            )
-            score -= 3  # type: ignore[operator]
-        else:
-            findings.append("- subprocess with shell=True: Not found (Good)")
-
-        # Check for potential secrets
-        secrets_pattern = r"api_key|password|secret|token"
-        secrets_found = grep_in_files(secrets_pattern, python_files)
-        if secrets_found > 0:
-            findings.append(
-                f"- Warning: Potential secrets found in {secrets_found} files (vars)"
-            )
-            score -= 1  # type: ignore[operator]
-
-        # Check for eval/exec
-        unsafe_eval = grep_in_files(r"eval\(|exec\(", python_files)
-        if unsafe_eval > 0:
-            findings.append(
-                f"- Critical: Unsafe eval/exec detected in {unsafe_eval} files"
-            )
-            score -= 3  # type: ignore[operator]
-
-    elif assessment_id == "G":  # Dependencies
-        has_req = Path("requirements.txt").exists()
-        has_pyproject = Path("pyproject.toml").exists()
-        findings.append(f"- requirements.txt: {'✓' if has_req else '✗'}")
-        findings.append(f"- pyproject.toml: {'✓' if has_pyproject else '✗'}")
-
-        # Check for lock files
-        has_poetry_lock = Path("poetry.lock").exists()
-        has_package_lock = Path("package-lock.json").exists()
-        has_req_lock = Path("requirements.lock").exists()
-        has_lock = has_poetry_lock or has_package_lock or has_req_lock
-        findings.append(
-            f"- Lock files (poetry.lock/package-lock.json/requirements.lock): "
-            f"{'✓' if has_lock else '✗'}"
-        )
-
-        if not (has_req or has_pyproject):
-            score -= 5  # type: ignore[operator]
-            findings.append("- Critical: No dependency definition found")
-
-        if not has_lock:
-            findings.append("- Note: No lock file found (reproducibility risk)")
-            score -= 1  # type: ignore[operator]
-
-    elif assessment_id == "H":  # CI/CD
-        has_github_workflows = Path(".github/workflows").exists()
-        findings.append(f"- GitHub Workflows: {'✓' if has_github_workflows else '✗'}")
-        if not has_github_workflows:
-            score -= 3  # type: ignore[operator]
-            findings.append("- Warning: No GitHub Actions workflows found")
-        else:
-            workflow_count = len(list(Path(".github/workflows").glob("*.yml"))) + len(
-                list(Path(".github/workflows").glob("*.yaml"))
-            )
-            findings.append(f"- Number of workflows: {workflow_count}")
-
-    elif assessment_id == "I":  # Code Style
-        ruff_result = run_ruff_check_wrapper()
-        black_result = run_black_check_wrapper()
-        ruff_status = "✓ passed" if ruff_result["exit_code"] == 0 else "✗ issues found"
-        findings.append(f"- Ruff check: {ruff_status}")
-        black_status = (
-            "✓ formatted" if black_result["exit_code"] == 0 else "✗ needs formatting"
-        )
-        findings.append(f"- Black formatting: {black_status}")
-        if ruff_result["exit_code"] != 0:
-            score -= 3  # type: ignore[operator]
-        if black_result["exit_code"] != 0:
-            score -= 2  # type: ignore[operator]
-
-    elif assessment_id == "J":  # API Design
-        # Check for type hints as a proxy for explicit API design
-        type_hint_pattern = r"def\s+.*\s*->\s*.*:"
-        files_with_hints = grep_in_files(type_hint_pattern, python_files)
-
-        findings.append(f"- Files with type hints: {files_with_hints}/{file_count}")
-
-        if file_count > 0:
-            ratio = files_with_hints / file_count
-            findings.append(f"- Type hint coverage: {ratio:.0%}")
-
-            if ratio < 0.2:
-                score -= 3  # type: ignore[operator]
-                findings.append("- Warning: Low type hint usage (< 20%)")
-            elif ratio < 0.5:
-                score -= 1  # type: ignore[operator]
-                findings.append("- Note: Moderate type hint usage")
-            else:
-                findings.append("- Good type hint usage")
-
-        # Check for abstract base classes
-        abc_usage = grep_in_files(
-            r"from abc import |class .*\(ABC\):|class .*\(Protocol\):", python_files
-        )
-        findings.append(f"- Files using ABC/Protocol: {abc_usage}")
-        if abc_usage > 0:
-            findings.append("- Detected use of Abstract Base Classes/Protocols")
-
-    elif assessment_id == "K":  # Data Handling
-        # Check for common data handling libraries
-        data_imports = grep_in_files(
-            r"import json|import pickle|import sqlite3|import csv|import pandas|"
-            r"import dataclasses",
-            python_files,
-        )
-        findings.append(f"- Files with data handling imports: {data_imports}")
-
-        # Check for data class usage
-        dataclass_usage = grep_in_files(r"@dataclass", python_files)
-        findings.append(f"- Files using @dataclass: {dataclass_usage}")
-
-        if data_imports == 0 and dataclass_usage == 0 and file_count > 5:
-            score -= 2  # type: ignore[operator]
-            findings.append(
-                "- Note: Limited explicit data handling structures detected"
-            )
-
-    elif assessment_id == "L":  # Logging
-        logging_imports = grep_in_files(
-            r"import logging|from logging import", python_files
-        )
-        findings.append(f"- Files importing logging: {logging_imports}")
-        if logging_imports < 2 and file_count > 5:
-            score -= 2  # type: ignore[operator]
-            findings.append("- Warning: Sparse logging detected")
-
-    elif assessment_id == "M":  # Configuration
-        config_files = (
-            list(Path(".").glob("*.ini"))
-            + list(Path(".").glob("*.toml"))
-            + list(Path(".").glob("*.env"))
-        )
-        findings.append(f"- Config files found: {[f.name for f in config_files]}")
-        if not config_files:
-            score -= 1  # type: ignore[operator]
-            findings.append("- Note: No standard config files found in root")
-
-    elif assessment_id == "N":  # Scalability
-        # Check for concurrency patterns
-        concurrency_usage = grep_in_files(
-            r"import asyncio|import multiprocessing|import threading|concurrent"
-            r"\.futures",
-            python_files,
-        )
-        findings.append(f"- Files with concurrency imports: {concurrency_usage}")
-
-        # Check for caching
-        cache_usage = grep_in_files(r"lru_cache|@cache", python_files)
-        findings.append(f"- Files using caching (lru_cache/cache): {cache_usage}")
-
-        if concurrency_usage == 0 and cache_usage == 0:
-            findings.append(
-                "- Note: No explicit concurrency or caching detected "
-                "(may affect scalability)"
-            )
-            # Neutral score impact unless context implies high scale needed
-
-    elif assessment_id == "O":  # Maintainability
-        # Simple maintainability metrics
-        long_files = 0
-        total_lines = 0
-
-        for p in python_files:
-            try:
-                line_count = len(p.read_text(encoding="utf-8").splitlines())
-                total_lines += line_count
-                if line_count > 1500:
-                    long_files += 1
-            except Exception:  # noqa: BLE001  # noqa: BLE001
-                pass
-
-        avg_lines = total_lines / file_count if file_count > 0 else 0
-        findings.append(f"- Average lines per file: {avg_lines:.0f}")
-        findings.append(f"- Files > 1500 lines: {long_files}")
-
-        if long_files > 0:
-            score -= min(3, long_files)  # type: ignore[operator]
-            findings.append(
-                f"- Warning: {long_files} large files detected (> 1500 lines)"
-            )
-
-        if avg_lines > 500:
-            score -= 1  # type: ignore[operator]
-            findings.append("- Warning: High average file size")
-
-    else:
-        # Fallback for unexpected assessment IDs
-        score = None
-        findings.append(f"- Python files analyzed: {file_count}")
-        findings.append(
-            "- **REQUIRES REVIEW**: No automated checks available for this category"
-        )
-        findings.append("- Score must be assigned by Jules bot or manual code review")
-
-    return findings, score
-
-
 def calculate_scores(score: int | None) -> tuple[int | None, str]:
-    """Clamp and format a raw score value.
-
-    Args:
-        score: Raw integer score (may be negative after deductions) or None
-               when the score cannot be determined automatically.
-
-    Returns:
-        A tuple of (clamped_score, score_display) where clamped_score is
-        clamped to [0, 10] (or None) and score_display is the formatted
-        string (e.g. "7/10" or "PENDING REVIEW").
-    """
+    """Clamp and format a raw score value."""
     if score is not None:
         clamped = max(0, min(10, score))
         return clamped, f"{clamped}/10"
@@ -515,15 +84,7 @@ def generate_report(
     findings: list[str],
     output_path: Path,
 ) -> None:
-    """Write the assessment markdown report to disk.
-
-    Args:
-        assessment_id:  Single uppercase letter identifying the assessment.
-        assessment:     Dict with 'name' and 'description' keys from ASSESSMENTS.
-        score_display:  Formatted score string (e.g. "7/10" or "PENDING REVIEW").
-        findings:       List of markdown bullet strings produced by analyze_module().
-        output_path:    Destination Path for the written report file.
-    """
+    """Write the assessment markdown report to disk."""
     report_content = f"""# Assessment {assessment_id}: {assessment["name"]}
 
 **Date**: {datetime.now().strftime("%Y-%m-%d %H:%M:%S")}
@@ -540,7 +101,7 @@ def generate_report(
 ## Recommendations
 
 - Review findings above
-- Address any ✗ items
+- Address any items marked N
 - Re-run assessment after fixes
 
 ## Automation Notes
@@ -550,22 +111,12 @@ This assessment was generated automatically. For detailed analysis:
 2. Review code manually for context-specific issues
 3. Create GitHub issues for actionable items
 """
-
-    # Ensure output directory exists
     output_path.parent.mkdir(parents=True, exist_ok=True)
-
-    # Write report
-    with open(output_path, "w", encoding="utf-8") as f:
-        f.write(report_content)
+    output_path.write_text(report_content, encoding="utf-8")
 
 
 def run_assessment(assessment_id: str, output_path: Path) -> int:
-    """
-    Run a specific assessment and generate report.
-
-    Thin orchestrator: delegates file discovery, analysis, scoring, and report
-    generation to the dedicated helper functions (file_discovery,
-    analyze_module, calculate_scores, generate_report).
+    """Run a specific assessment and generate report.
 
     Args:
         assessment_id: Assessment ID (A-O)
@@ -576,33 +127,21 @@ def run_assessment(assessment_id: str, output_path: Path) -> int:
     """
     assessment = ASSESSMENTS.get(assessment_id)
     if not assessment:
-        logger.error(f"Unknown assessment: {assessment_id}")
+        logger.error("Unknown assessment: %s", assessment_id)
         return 1
 
-    logger.info(f"Running Assessment {assessment_id}: {assessment['name']}...")
-
-    # DbC precondition: assessment must be valid before proceeding
-    if not (assessment is not None):
-        raise ValueError("DbC Blocked: Precondition failed.")
-
-    # 1. Discover files
+    logger.info("Running Assessment %s: %s...", assessment_id, assessment["name"])
     python_files, file_count = file_discovery()
-
-    # 2. Analyse
     findings, raw_score = analyze_module(assessment_id, python_files, file_count)
-
-    # 3. Score
     _, score_display = calculate_scores(raw_score)
-
-    # 4. Report
     generate_report(assessment_id, assessment, score_display, findings, output_path)
-
-    logger.info(f"✓ Assessment {assessment_id} report saved to {output_path}")
-    logger.info(f"  Score: {score_display}")
+    logger.info("Assessment %s report saved to %s", assessment_id, output_path)
+    logger.info("  Score: %s", score_display)
     return 0
 
 
-def main():
+def main() -> None:
+    """Entry point."""
     parser = argparse.ArgumentParser(description="Run repository assessment")
     parser.add_argument(
         "--assessment",
@@ -616,11 +155,8 @@ def main():
         type=Path,
         help="Output file path for assessment report",
     )
-
     args = parser.parse_args()
-
-    exit_code = run_assessment(args.assessment, args.output)
-    sys.exit(exit_code)
+    sys.exit(run_assessment(args.assessment, args.output))
 
 
 if __name__ == "__main__":
