+++ src/games/shared/combat_manager.py
@@ -1,7 +1,8 @@
 """Base class for combat managers.
 
-Encapsulates hit detection, damage application, kill tracking,
-and explosion logic that was previously duplicated across game.py files.
+Encapsulates hit detection, kill tracking, and hitscan shot logic.
+Explosion helpers live in combat_explosions.py (_CombatExplosionsMixin).
+Shot visual helpers live in combat_visuals.py (_CombatVisualsMixin).
 
 Subclasses override constants and game-specific behavior as needed.
 """
@@ -16,8 +17,8 @@ from contextlib import suppress
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
-import pygame
-
+from games.shared.combat_explosions import _CombatExplosionsMixin
+from games.shared.combat_visuals import _CombatVisualsMixin
 from games.shared.constants import COMBO_TIMER_FRAMES
 
 if TYPE_CHECKING:
@@ -51,7 +52,7 @@ class ShotResolutionResult:
     is_headshot: bool
 
 
-class CombatManagerBase:
+class CombatManagerBase(_CombatExplosionsMixin, _CombatVisualsMixin):
     """Manages hit detection, damage calculation, and explosions.
 
     Dependencies are injected via the constructor -- the manager never
@@ -76,26 +77,14 @@ class CombatManagerBase:
         self.particle_system = particle_system
         self.sound_manager = sound_manager
         self.C = constants
-
-        # Optional callback invoked after each kill: on_kill(bot)
         self._on_kill = on_kill
-
-        # Kill tracking state -- the game reads these back
         self.kills = 0
         self.kill_combo_count = 0
         self.kill_combo_timer = 0
         self.last_death_pos: tuple[float, float] | None = None
 
-    # ------------------------------------------------------------------
-    # Kill handling
-    # ------------------------------------------------------------------
-
     def handle_kill(self, bot: Any) -> None:
-        """Record a kill and update combo state.
-
-        Args:
-            bot: The bot that was killed.
-        """
+        """Record a kill and update combo state."""
         self.kills += 1
         self.kill_combo_count += 1
         self.kill_combo_timer = COMBO_TIMER_FRAMES
@@ -105,46 +94,12 @@ class CombatManagerBase:
         if self._on_kill is not None:
             self._on_kill(bot)
 
-    # ------------------------------------------------------------------
-    # Hitscan shot detection
-    # ------------------------------------------------------------------
-
-    def _apply_primary_shot(
-        self, request: ShotResolutionRequest, resolution: Any, weapon_range: float
-    ) -> None:
-        """Apply laser, damage, and visuals for a primary (non-secondary) shot."""
-        if request.is_laser:
-            self.particle_system.add_laser(
-                start=(self.C.SCREEN_WIDTH - 200, self.C.SCREEN_HEIGHT - 180),
-                end=(self.C.SCREEN_WIDTH // 2, self.C.SCREEN_HEIGHT // 2),
-                color=(255, 0, 0),
-                timer=5,
-                width=3,
-            )
-        if resolution.closest_bot:
-            self._apply_hit_damage(
-                request.player,
-                resolution.closest_bot,
-                resolution.closest_distance,
-                weapon_range,
-                request.player.get_current_weapon_damage(),
-                resolution.is_headshot,
-                request.damage_texts,
-                request.show_damage,
-            )
-        self._add_shot_visuals(
-            request.player,
-            resolution.aim_angle,
-            resolution.closest_bot,
-            resolution.closest_distance,
-            resolution.wall_distance,
-        )
-
     def check_shot_hit(self, request: ShotResolutionRequest) -> list[dict[str, Any]]:
         """Resolve a hitscan shot and apply the shared side effects."""
         try:
             resolution = self.resolve_shot_target(request)
             weapon_range = self._get_weapon_range(request.player, request.is_secondary)
+
             if request.is_secondary:
                 self._handle_secondary_hit(
                     request.player,
@@ -154,15 +109,44 @@ class CombatManagerBase:
                     request.damage_texts,
                 )
                 return request.damage_texts
-            self._apply_primary_shot(request, resolution, weapon_range)
+
+            if request.is_laser:
+                self.particle_system.add_laser(
+                    start=(self.C.SCREEN_WIDTH - 200, self.C.SCREEN_HEIGHT - 180),
+                    end=(self.C.SCREEN_WIDTH // 2, self.C.SCREEN_HEIGHT // 2),
+                    color=(255, 0, 0),
+                    timer=5,
+                    width=3,
+                )
+
+            if resolution.closest_bot:
+                self._apply_hit_damage(
+                    request.player,
+                    resolution.closest_bot,
+                    resolution.closest_distance,
+                    weapon_range,
+                    request.player.get_current_weapon_damage(),
+                    resolution.is_headshot,
+                    request.damage_texts,
+                    request.show_damage,
+                )
+
+            self._add_shot_visuals(
+                request.player,
+                resolution.aim_angle,
+                resolution.closest_bot,
+                resolution.closest_distance,
+                resolution.wall_distance,
+            )
         except (ValueError, AttributeError, ZeroDivisionError):
             logger.exception("Error in check_shot_hit")
+
         return request.damage_texts
 
     def resolve_shot_target(
         self, request: ShotResolutionRequest
     ) -> ShotResolutionResult:
-        """Resolve hit geometry without applying presentation or damage side effects."""
+        """Resolve hit geometry without side effects."""
         weapon_range = self._get_weapon_range(request.player, request.is_secondary)
         aim_angle = self._resolve_aim_angle(request.player, request.angle_offset)
         wall_distance = self._resolve_wall_distance(
@@ -194,8 +178,7 @@ class CombatManagerBase:
         spread_base = getattr(self.C, "SPREAD_BASE", 0.04)
         current_spread = spread_zoom if player.zoomed else spread_base
         if angle_offset == 0.0:
-            spread_offset = random.uniform(-current_spread, current_spread)
-            angle = player.angle + spread_offset
+            angle = player.angle + random.uniform(-current_spread, current_spread)
         else:
             angle = player.angle + angle_offset
         return angle % (2 * math.pi)
@@ -228,405 +211,21 @@ class CombatManagerBase:
         for bot in bots:
             if not bot.alive:
                 continue
-
             dx = bot.x - player.x
             dy = bot.y - player.y
             dist_sq = dx * dx + dy * dy
             if dist_sq > wall_dist_sq:
                 continue
-
             distance = math.sqrt(dist_sq)
             bot_angle = math.atan2(dy, dx)
             if bot_angle < 0:
                 bot_angle += 2 * math.pi
-
             angle_diff = abs(bot_angle - aim_angle)
             if angle_diff > math.pi:
                 angle_diff = 2 * math.pi - angle_diff
-
             if angle_diff < 0.15 and distance < closest_dist:
                 closest_bot = bot
                 closest_dist = distance
                 is_headshot = angle_diff < headshot_threshold
 
         return closest_bot, closest_dist, is_headshot
-
-    def _handle_secondary_hit(
-        self,
-        player: Any,
-        closest_bot: Any | None,
-        closest_dist: float,
-        wall_dist: float,
-        damage_texts: list[dict[str, Any]],
-    ) -> None:
-        """Handle secondary fire (laser explosion at impact)."""
-        impact_dist = wall_dist
-        if closest_bot and closest_dist < wall_dist:
-            impact_dist = closest_dist
-
-        ray_angle = player.angle
-        impact_x = player.x + math.cos(ray_angle) * impact_dist
-        impact_y = player.y + math.sin(ray_angle) * impact_dist
-
-        try:
-            self.explode_laser(impact_x, impact_y, damage_texts)
-        except (ValueError, AttributeError, pygame.error):
-            logger.exception("Error in explode_laser")
-
-        laser_duration = getattr(self.C, "LASER_DURATION", 10)
-        laser_width = getattr(self.C, "LASER_WIDTH", 5)
-        self.particle_system.add_laser(
-            start=(self.C.SCREEN_WIDTH - 200, self.C.SCREEN_HEIGHT - 180),
-            end=(self.C.SCREEN_WIDTH // 2, self.C.SCREEN_HEIGHT // 2),
-            color=(0, 255, 255),
-            timer=laser_duration,
-            width=laser_width,
-        )
-
-    def _apply_hit_damage(
-        self,
-        player: Any,
-        bot: Any,
-        distance: float,
-        weapon_range: float,
-        base_damage: int,
-        is_headshot: bool,
-        damage_texts: list[dict[str, Any]],
-        show_damage: bool,
-    ) -> None:
-        """Apply damage to a bot that was hit."""
-        final_damage = self._calculate_final_damage(
-            player, bot, distance, weapon_range, base_damage
-        )
-        bot.take_damage(final_damage, is_headshot=is_headshot)
-        if show_damage:
-            self._append_damage_text(damage_texts, final_damage, is_headshot)
-        self._spawn_hit_particles()
-        if not bot.alive:
-            self.handle_kill(bot)
-
-    def _calculate_final_damage(
-        self,
-        player: Any,
-        bot: Any,
-        distance: float,
-        weapon_range: float,
-        base_damage: int,
-    ) -> int:
-        """Return damage after applying range falloff and accuracy scaling."""
-        range_factor = max(0.3, 1.0 - (distance / weapon_range))
-        dx = bot.x - player.x
-        dy = bot.y - player.y
-        bot_angle = math.atan2(dy, dx)
-        if bot_angle < 0:
-            bot_angle += 2 * math.pi
-        angle_diff = abs(bot_angle - player.angle)
-        if angle_diff > math.pi:
-            angle_diff = 2 * math.pi - angle_diff
-        accuracy_factor = max(0.5, 1.0 - (angle_diff / 0.15))
-        return int(base_damage * range_factor * accuracy_factor)
-
-    def _append_damage_text(
-        self,
-        damage_texts: list[dict[str, Any]],
-        final_damage: int,
-        is_headshot: bool,
-    ) -> None:
-        """Append a floating damage number to the damage_texts list."""
-        damage_color = getattr(self.C, "DAMAGE_TEXT_COLOR", (255, 255, 255))
-        damage_texts.append(
-            {
-                "x": self.C.SCREEN_WIDTH // 2 + random.randint(-20, 20),
-                "y": self.C.SCREEN_HEIGHT // 2 - 50,
-                "text": str(final_damage) + ("!" if is_headshot else ""),
-                "color": self.C.RED if is_headshot else damage_color,
-                "timer": 60,
-                "vy": -1.0,
-            }
-        )
-
-    def _spawn_hit_particles(self) -> None:
-        """Spawn the screen-space explosion and blood-splash particles."""
-        sw2 = self.C.SCREEN_WIDTH // 2
-        sh2 = self.C.SCREEN_HEIGHT // 2
-        self.particle_system.add_explosion(sw2, sh2, count=5)
-        blood_color = getattr(self.C, "BLUE_BLOOD", (0, 0, 200))
-        particle_lifetime = getattr(self.C, "PARTICLE_LIFETIME", 30)
-        for _ in range(10):
-            self.particle_system.add_particle(
-                x=sw2,
-                y=sh2,
-                dx=random.uniform(-5, 5),
-                dy=random.uniform(-5, 5),
-                color=blood_color,
-                timer=particle_lifetime,
-                size=random.randint(2, 5),
-            )
-
-    def _add_wall_hit_particles(
-        self,
-        player: Any,
-        aim_angle: float,
-        wall_dist: float,
-        screen_hit_x: float,
-        screen_hit_y: float,
-    ) -> None:
-        """Spawn wall-impact explosion and screen sparks for a missed shot."""
-        hit_world_x = player.x + math.cos(aim_angle) * wall_dist
-        hit_world_y = player.y + math.sin(aim_angle) * wall_dist
-        self.particle_system.add_world_explosion(
-            hit_world_x, hit_world_y, 0.5, count=10, color=(255, 200, 100), speed=0.1
-        )
-        for _ in range(5):
-            self.particle_system.add_particle(
-                x=screen_hit_x,
-                y=screen_hit_y,
-                dx=random.uniform(-3, 3),
-                dy=random.uniform(-3, 3),
-                color=(255, 200, 100),
-                timer=20,
-                size=random.randint(1, 3),
-            )
-
-    def _add_shot_visuals(
-        self,
-        player: Any,
-        aim_angle: float,
-        closest_bot: Any | None,
-        closest_dist: float,
-        wall_dist: float,
-    ) -> None:
-        """Add tracer and impact visual effects."""
-        fov = getattr(self.C, "FOV", 1.0)
-        angle_diff = (aim_angle - player.angle + math.pi) % (2 * math.pi) - math.pi
-        screen_hit_x = (0.5 - angle_diff / fov) * self.C.SCREEN_WIDTH
-        screen_hit_y = self.C.SCREEN_HEIGHT // 2 + getattr(player, "pitch", 0)
-        self.particle_system.add_trace(
-            start=(self.C.SCREEN_WIDTH * 0.6, self.C.SCREEN_HEIGHT),
-            end=(screen_hit_x, screen_hit_y),
-            color=(255, 255, 100),
-            timer=5,
-            width=1,
-        )
-        if not closest_bot:
-            self._add_wall_hit_particles(
-                player, aim_angle, wall_dist, screen_hit_x, screen_hit_y
-            )
-
-    # ------------------------------------------------------------------
-    # Explosion helpers
-    # ------------------------------------------------------------------
-
-    def handle_bomb_explosion(
-        self,
-        player: Any,
-        bots: list[Any],
-        damage_texts: list[dict[str, Any]],
-    ) -> list[dict[str, Any]]:
-        """Handle bomb explosion centered on the player.
-
-        Returns the updated damage_texts list.
-        """
-        sw2 = self.C.SCREEN_WIDTH // 2
-        sh2 = self.C.SCREEN_HEIGHT // 2
-        self._spawn_bomb_particles(player, sw2, sh2)
-        self._apply_bomb_damage(player, bots, sw2, sh2)
-        try:
-            self.sound_manager.play_sound("bomb")
-        except (pygame.error, OSError):
-            logger.exception("Bomb Audio Failed")
-        damage_texts.append(
-            {
-                "x": sw2,
-                "y": sh2 - 100,
-                "text": "BOMB DROPPED!",
-                "color": self.C.ORANGE,
-                "timer": 120,
-                "vy": -0.5,
-            }
-        )
-        return damage_texts
-
-    def _spawn_bomb_particles(self, player: Any, sw2: int, sh2: int) -> None:
-        """Spawn the flash particle, world explosion, and radial debris."""
-        self.particle_system.add_particle(
-            x=sw2, y=sh2, dx=0, dy=0, color=self.C.WHITE, timer=40, size=3000
-        )
-        self.particle_system.add_world_explosion(
-            player.x, player.y, 0.5, count=100, color=self.C.ORANGE, speed=0.5
-        )
-        dark_red = getattr(self.C, "DARK_RED", (139, 0, 0))
-        colors = [self.C.ORANGE, self.C.RED, self.C.YELLOW, dark_red, (50, 50, 50)]
-        for _ in range(300):
-            angle = random.uniform(0, 2 * math.pi)
-            speed = random.uniform(5, 25)
-            self.particle_system.add_particle(
-                x=sw2,
-                y=sh2,
-                dx=math.cos(angle) * speed,
-                dy=math.sin(angle) * speed,
-                color=random.choice(colors),
-                timer=random.randint(40, 100),
-                size=random.randint(5, 25),
-            )
-
-    def _apply_bomb_damage(
-        self, player: Any, bots: list[Any], sw2: int, sh2: int
-    ) -> None:
-        """Deal 1000 damage to all bots within BOMB_RADIUS of the player."""
-        bomb_radius = getattr(self.C, "BOMB_RADIUS", 10.0)
-        for bot in bots:
-            if not bot.alive:
-                continue
-            dx = bot.x - player.x
-            dy = bot.y - player.y
-            dist = math.sqrt(dx * dx + dy * dy)
-            if dist < bomb_radius:
-                if bot.take_damage(1000):
-                    self.handle_kill(bot)
-                if dist < 5.0:
-                    self.particle_system.add_explosion(sw2, sh2, count=3)
-
-    def explode_laser(
-        self,
-        impact_x: float,
-        impact_y: float,
-        damage_texts: list[dict[str, Any]],
-    ) -> list[dict[str, Any]]:
-        """Trigger a laser AOE explosion at an impact point."""
-        try:
-            self.sound_manager.play_sound("boom_real")
-        except (pygame.error, OSError):
-            logger.exception("Boom sound failed")
-        try:
-            sw2 = self.C.SCREEN_WIDTH // 2
-            sh2 = self.C.SCREEN_HEIGHT // 2
-            hits = self._apply_laser_damage(impact_x, impact_y, sw2, sh2)
-            if hits > 0:
-                damage_texts.append(
-                    {
-                        "x": sw2,
-                        "y": sh2 - 80,
-                        "text": "LASER ANNIHILATION!",
-                        "color": (255, 0, 255),
-                        "timer": 60,
-                        "vy": -2,
-                    }
-                )
-        except (ValueError, AttributeError, pygame.error):
-            logger.exception("Critical Laser Error")
-        return damage_texts
-
-    def _apply_laser_damage(
-        self, impact_x: float, impact_y: float, sw2: int, sh2: int
-    ) -> int:
-        """Spawn laser explosion particles and damage bots in AOE radius.
-
-        Returns the number of bots hit.
-        """
-        self.particle_system.add_world_explosion(
-            impact_x, impact_y, 0.5, count=80, color=(0, 255, 255), speed=0.4
-        )
-        laser_radius = getattr(self.C, "LASER_AOE_RADIUS", 5.0)
-        hits = 0
-        for bot in self.entity_manager.bots:
-            if not getattr(bot, "alive", False):
-                continue
-            dist = math.sqrt((bot.x - impact_x) ** 2 + (bot.y - impact_y) ** 2)
-            if dist < laser_radius:
-                killed = bot.take_damage(500)
-                hits += 1
-                if killed:
-                    self.handle_kill(bot)
-                for _ in range(10):
-                    self.particle_system.add_particle(
-                        x=sw2,
-                        y=sh2,
-                        dx=random.uniform(-10, 10),
-                        dy=random.uniform(-10, 10),
-                        color=(
-                            random.randint(200, 255),
-                            0,
-                            random.randint(200, 255),
-                        ),
-                        timer=40,
-                        size=random.randint(4, 8),
-                    )
-        return hits
-
-    def explode_generic(
-        self,
-        projectile: Any,
-        radius: float,
-        weapon_type: str,
-        player: Any,
-        damage_flash_timer: int,
-    ) -> int:
-        """Generic AOE explosion around a projectile impact.
-
-        Returns the updated damage_flash_timer value.
-        """
-        dist_to_player = math.sqrt(
-            (projectile.x - player.x) ** 2 + (projectile.y - player.y) ** 2
-        )
-        if dist_to_player < 15:
-            damage_flash_timer = 15
-        sw2 = self.C.SCREEN_WIDTH // 2
-        sh2 = self.C.SCREEN_HEIGHT // 2
-        self._spawn_generic_explosion_fx(
-            projectile, weapon_type, dist_to_player, sw2, sh2
-        )
-        self._apply_generic_aoe_damage(projectile, radius)
-        return damage_flash_timer
-
-    def _spawn_generic_explosion_fx(
-        self,
-        projectile: Any,
-        weapon_type: str,
-        dist_to_player: float,
-        sw2: int,
-        sh2: int,
-    ) -> None:
-        """Spawn world explosion particles and play the impact sound."""
-        color = self.C.CYAN if weapon_type == "plasma" else self.C.ORANGE
-        self.particle_system.add_world_explosion(
-            projectile.x,
-            projectile.y,
-            getattr(projectile, "z", 0.5),
-            count=50,
-            color=color,
-            speed=0.3,
-        )
-        if dist_to_player < 20:
-            count = 5 if weapon_type == "rocket" else 3
-            self.particle_system.add_explosion(sw2, sh2, count=count)
-            for _ in range(20):
-                self.particle_system.add_particle(
-                    x=sw2,
-                    y=sh2,
-                    dx=random.uniform(-10, 10),
-                    dy=random.uniform(-10, 10),
-                    color=color,
-                    timer=40,
-                    size=random.randint(4, 8),
-                )
-        try:
-            self.sound_manager.play_sound(
-                "boom_real" if weapon_type == "rocket" else "shoot_plasma"
-            )
-        except (pygame.error, OSError):
-            pass
-
-    def _apply_generic_aoe_damage(self, projectile: Any, radius: float) -> None:
-        """Apply distance-scaled damage to all bots within the explosion radius."""
-        for bot in self.entity_manager.bots:
-            if not bot.alive:
-                continue
-            dx = bot.x - projectile.x
-            dy = bot.y - projectile.y
-            dist = math.sqrt(dx * dx + dy * dy)
-            if dist < radius:
-                damage_factor = 1.0 - (dist / radius)
-                damage = int(projectile.damage * damage_factor)
-                if bot.take_damage(damage):
-                    self.handle_kill(bot)
